Rewrote the Override Margins handling for tables
The override margins section for row/cell will now properly reflect the
current margins for the given selection, and the override checkbox
setting will try to reflect the current state of the row/cell. The one
caveat is that if a cell is checked as “override”, but the margins
match the row/table, it will not show up as checked the next time the
cell properties are opened.