Fixed a bug with padding not being properly reported
We kept track of whether a row/cell had overridden the table padding by
placing the default padding on the table element. However, in FireFox
using the css call failed even when the table element contained the
style. So instead, we are calling style to ensure the proper value is
returned.