Fixed two issues with drag and drop image files in IE
1. If a field is set as read-only prior to gaining focus, then the
draggingInternally flag is improperly set on mousedown, with no way of
being unset because there will never be a dragend event. This prevents
dropping data on a field from an external source after clicking on it.
2. getDataTransferItems was not written to support IE. IE DataTransfer
objects do NOT support anything but 'Text' or 'URL' passed into the
getData function; and will throw an exception if you do. This was
breaking dropping external image files on IE, because the drop would be
killed by this exception. I added proper checking to this function to
ensure we are properly supporting IE.