Fixed functional issues with table alignment and background color
We were using obsolete/deprecated HTML attributes, instead of using CSS
to style the table and its contents. Because some of the logic was
checking/using CSS, while others were checking/using the attributes,
this led to some inconsistencies in behavior, and cases where setting
the attribute would fail because the CSS had already been set.