Removing dashed border on hover for no table borders
While this used to work, apparently it no longer does after fixing other
defects. We can look at making this function properly at a later point
(perhaps we need to manually apply this when borders are set instead?),
but for now it's not worth having this since it no longer functions.