Adjusted quirks webkit custom delete
In webkit there is a custom delete that is used which results in
removing spans that were inserted as part of the delete. These spans can
contain styling information that is lost. Because of this the spans are
now kept but the 65279 character is now removed if inserted by the
browsers behavior at the start of the span.