Cache Styles for Copy, Cut, & Paste
Added logic that will cache the styles of the selected elements when doing
a copy or paste operation and adjusted logic on paste for WebKit to prefer
the cached styles over the style attribute. The net result of this is that
as much as possible copying/cutting in an editor and then pasting in an
editor will always retain all styles.

Also tweaked the build script to cleanup after itself after a build.