Fixed delist of first list item
If the first list item has sub items and is delisted the sub items
create a new list instead of being correctly merged with the remaining
items in the list. This previously would result in a new list with one
item (the first sub item) and the rest of the old list directly below.