Removed blob caching
Adjusted image handling to not use blob caching since we replace the image immediately anyway. Additionally, this prevents an error where createObjectURL in chrome loses the type when this is called with a FILE object instead of a BLOB, this would lead to an invalid image type error when the blob is then retrieved.