isWordContent includes HTML with scripts
Adjusted isWordContent check to also flag content that contains scripts. This will mean the content goes through the additional processing used for when the HTML content is from Word, Office 365, and Google docs. This will ensure the scripts are removed and put in additional procesisng since this is likely a paste of a full webpage/iframe.