Revert delay for showing context toolbar to 0
When showing an Angular dialog for SproutCore WYSIWYG the context
toolbar was being shown on top of the Angular dialog. This was because
lazy loaded Angular dialogs show a loading dialog first while lazy
loading the component. The SproutCore view briefly regains focus when
the loading dialog is closed which attempts to reshow the context
toolbar for the selected content. The delay in showing the context
toolbar mean that this was resolved after the actual Angular dialog was
shown.